Start per-node servers regardless of feature flag
Previously rexi_server_mon would fail to start per-node servers on a
node where the use of those servers in rexi:cast messages was disabled
by the "server_per_node" feature flag.  This mostly defeats the purpose
of the feature flag, which is to ensure that we can start all the
per-node servers before we try to use them.  Moreover, it cluttered up
the error logs because rexi_server_mon would crash after trying to
execute a `start_child` on rexi_server.

The fix is not DRY but it does the job well.

BugzID: 18970