Cancel late-arriving streaming workers
This patch causes rexi's receive loop to cancel any workers that try to
initialize a stream when they're not part of the current worker set.
Previously we used to ignore them and let them timeout.

The patch is not bulletproof; a worker could initialize a stream after
the receive loop has completed.  If the process running the receive loop
does not enter another receive loop the worker will eventually timeout.
It's an intermediate improvement, though.

BugzID: 24507